creates an endpoint for fetching bookmarked repos
